requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0
requests>=2.28.2
django-redis>=5.2.0
redis>=4.5.4
celery>=5.3.0
django-celery-beat>=2.5.0